+++ src/react/hooks/useRepeaterControl.ts
@@ -5,6 +5,7 @@ import { UI_STRINGS } from '../../utils/constants';
 import { groupFormFields, type TFormFieldGroup } from '../../utils/formFieldGrouping';
 import { isFieldRequired, isFieldVisible } from '../../utils/formFieldHelpers';
 import { getRepeaterCountText } from '../../utils/repeaterCountText';
+import { resolveFormFieldDefaultValue } from '../../utils/formFieldDefaults';
 import {
   assignRepeaterItemId,
   isAutoRepeaterItemIdField,
@@ -99,29 +100,7 @@ export function useRepeaterControl({
       } else if (field.defaultValue !== undefined) {
         newItem[field.field] = field.defaultValue;
       } else {
-        switch (field.type) {
-          case 'checkbox':
-            newItem[field.field] = false;
-            break;
-          case 'number':
-            newItem[field.field] = 0;
-            break;
-          case 'api-select':
-            newItem[field.field] = field.apiSelectConfig?.multiple ? [] : null;
-            break;
-          case 'image':
-          case 'file':
-            newItem[field.field] = field.multiple ? [] : '';
-            break;
-          case 'repeater':
-            newItem[field.field] = [];
-            break;
-          case 'custom':
-            newItem[field.field] = '';
-            break;
-          default:
-            newItem[field.field] = '';
-        }
+        newItem[field.field] = resolveFormFieldDefaultValue(field);
       }
     });